Moved and renamed useSquareGrid to useSquareStore, broke out actionoble functionality into a new hook (useSquareActions), refactored the folder/file structure relating to the store into clear hirarcy with subfolders.
