+++ client/src/hooks/stores/useSquareStore/useSquareActions/partials/useAddSquare/utils/create/createNewSquare/utils/colors/generateSquareColor.ts
