+++ client/src/hooks/stores/useSquareStore/useSquareActions/partials/useAddSquare/utils/get/getPreviousSquare.ts
@@ -1,0 +1,7 @@
+import type { BaseSquareType } from "@models/types/square";
+
+const getPreviousSquare = (squares: BaseSquareType[]) => {
+	return squares.length > 0 ? squares[squares.length - 1] : null;
+};
+
+export { getPreviousSquare };
