+++ server/api/Data/LastSquareData.json
@@ -1,6 +1,6 @@
 {
-  "Id": 2,
-  "X": 1,
-  "Y": 0,
-  "Color": "#acfd85"
+  "Id": 16,
+  "X": 0,
+  "Y": 3,
+  "Color": "#84d19e"
 }
