+++ server/api/Data/SquareData.json
@@ -10,5 +10,89 @@
     "X": 1,
     "Y": 0,
     "Color": "#acfd85"
+  },
+  {
+    "Id": 3,
+    "X": 1,
+    "Y": 1,
+    "Color": "#4feaf8"
+  },
+  {
+    "Id": 4,
+    "X": 0,
+    "Y": 1,
+    "Color": "#9ec947"
+  },
+  {
+    "Id": 5,
+    "X": 2,
+    "Y": 0,
+    "Color": "#20a9cd"
+  },
+  {
+    "Id": 6,
+    "X": 2,
+    "Y": 1,
+    "Color": "#557dd5"
+  },
+  {
+    "Id": 7,
+    "X": 2,
+    "Y": 2,
+    "Color": "#ee4043"
+  },
+  {
+    "Id": 8,
+    "X": 1,
+    "Y": 2,
+    "Color": "#32d40c"
+  },
+  {
+    "Id": 9,
+    "X": 0,
+    "Y": 2,
+    "Color": "#a2c972"
+  },
+  {
+    "Id": 10,
+    "X": 3,
+    "Y": 0,
+    "Color": "#f28212"
+  },
+  {
+    "Id": 11,
+    "X": 3,
+    "Y": 1,
+    "Color": "#37ef7b"
+  },
+  {
+    "Id": 12,
+    "X": 3,
+    "Y": 2,
+    "Color": "#3209f1"
+  },
+  {
+    "Id": 13,
+    "X": 3,
+    "Y": 3,
+    "Color": "#8b551f"
+  },
+  {
+    "Id": 14,
+    "X": 2,
+    "Y": 3,
+    "Color": "#38342b"
+  },
+  {
+    "Id": 15,
+    "X": 1,
+    "Y": 3,
+    "Color": "#3031d6"
+  },
+  {
+    "Id": 16,
+    "X": 0,
+    "Y": 3,
+    "Color": "#84d19e"
   }
 ]
